Move assisted tagging mapping to assist item from popup
