+++ app/views/Project/EntryEdit/queries.ts
@@ -160,6 +160,13 @@ export const PROJECT_FRAMEWORK = gql`
                     width
                     version
                 }
+                predictionTagsMapping {
+                    id
+                    tag
+                    widget
+                    widgetType
+                    association
+                }
             }
         }
     }
